+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.7.1 - Unreleased
 ### Changed
+- Number of expected recipients is now calculated in the background on the edit sendout screen to prevent slow page load.
+- Improved how the `@web` alias is determined to be in the site or asset URLs in the preflight check.
 - Changed `Craft::warning` to `Craft::info` when logging user actions.
 
 ### Fixed